fix precision error in mpi test

+++ tests/internal/scripts/mpi/nonblocking.py
@@ -43,9 +43,9 @@ reqs = np.zeros(4, 'int32')
 before = np.int32(rank - 1)
 after  = np.int32(rank + 1)
 if rank == 0:
-    before = size - 1
+    before = np.int32(size - 1)
 if rank == size - 1:
-    after  = 0
+    after  = np.int32(0)
 
 # ...
 mpi_irecv(x, n, MPI_REAL8, before, tag0, comm, reqs[0], ierr)
